requirements.txt
google-api-python-client
google-cloud-language
